Creating UserMenu component
